Created car.py, lists.py, main.py, and player.py.
just did basic __init__ for player to be visible  on screen. there are no user controls yet.

main.py creates a screen loads the player and generates a number of cars randomly placed and which move across the screen in both directions.

lists contains list of colours for the car turtles and names for the car turtles for generation purposes.